Fixed Firefox and Chrome versions of Index + Products
Fixed header, footer, slider, rounded corners, nav and gradients for
index and product pages. Secret to finding the problem: there was
something wrong with the <div id="wrapper" AND when using style.css it
is very important that the styles occur somewhat in the order that they
appear when using the 99lime.com HTML-Kickstart CSS files.